+++ modules/base/util/util-collection-via-trove/src/main/java/consulo/util/collection/trove/impl/longs/MyLongHashSet.java
@@ -62,7 +62,7 @@ public class MyLongHashSet implements LongSet {
   }
 
   @Override
-  public boolean remove(int value) {
+  public boolean remove(long value) {
     return mySet.remove(value);
   }
